requirements.txt
bs4
faker
pyexecjs
requests
telethon